+++ features/aave/manage/containers/AaveManageTabBar.tsx
@@ -3,7 +3,7 @@ import { useAutomationContext } from 'components/AutomationContextProvider'
 import { TabBar } from 'components/TabBar'
 import { ProtectionControl } from 'components/vault/ProtectionControl'
 import { IStrategyConfig } from 'features/aave/common/StrategyConfigTypes'
-import { supportsAaveStopLoss } from 'features/aave/helpers/supportsStopLoss'
+import { supportsAaveStopLoss } from 'features/aave/helpers/supportsAaveStopLoss'
 import { useManageAaveStateMachineContext } from 'features/aave/manage/containers/AaveManageStateMachineContext'
 import { SidebarManageAaveVault } from 'features/aave/manage/sidebars/SidebarManageAaveVault'
 import { isSupportedAutomationTokenPair } from 'features/automation/common/helpers'
