+++ features/aave/open/state/openAaveStateMachine.ts
@@ -25,7 +25,7 @@ import {
   RefTransactionMachine,
 } from 'features/aave/common/BaseAaveContext'
 import { ProxyType } from 'features/aave/common/StrategyConfigTypes'
-import { supportsAaveStopLoss } from 'features/aave/helpers/supportsStopLoss'
+import { supportsAaveStopLoss } from 'features/aave/helpers/supportsAaveStopLoss'
 import {
   AutomationAddTriggerData,
   AutomationAddTriggerTxDef,
